Separates extractor for each S-expression version
massign structure patterns changed for Ruby 2.5 or later.

```
$ docker run -it --rm rubylang/all-ruby env ALL_RUBY_SINCE=ruby-2.4 ./all-ruby -rpp -rripper -e 'pp Ripper.sexp("a, (b, c) = foo")'
ruby-2.4.0          [:program,
                     [[:massign,
                       [[:@ident, "a", [1, 0]],
                        [:mlhs_paren, [[:@ident, "b", [1, 4]], [:@ident, "c", [1, 7]]]]],
                       [:vcall, [:@ident, "foo", [1, 12]]]]]]
...
ruby-2.4.9          [:program,
                     [[:massign,
                       [[:@ident, "a", [1, 0]],
                        [:mlhs_paren, [[:@ident, "b", [1, 4]], [:@ident, "c", [1, 7]]]]],
                       [:vcall, [:@ident, "foo", [1, 12]]]]]]
ruby-2.5.0-preview1 [:program,
                     [[:massign,
                       [[:var_field, [:@ident, "a", [1, 0]]],
                        [:mlhs,
                         [:var_field, [:@ident, "b", [1, 4]]],
                         [:var_field, [:@ident, "c", [1, 7]]]]],
                       [:vcall, [:@ident, "foo", [1, 12]]]]]]
...
ruby-2.7.0-rc2      [:program,
                     [[:massign,
                       [[:var_field, [:@ident, "a", [1, 0]]],
                        [:mlhs,
                         [:var_field, [:@ident, "b", [1, 4]]],
                         [:var_field, [:@ident, "c", [1, 7]]]]],
                       [:vcall, [:@ident, "foo", [1, 12]]]]]]
```

```
$ docker run -it --rm rubylang/all-ruby env ALL_RUBY_SINCE=ruby-2.4 ./all-ruby -rpp -rripper -e 'pp Ripper.sexp("a,*b=1,2")'

ruby-2.4.0          [:program,
                     [[:massign,
                       [:mlhs_add_star, [[:@ident, "a", [1, 0]]], [:@ident, "b", [1, 3]]],
                       [:mrhs_new_from_args, [[:@int, "1", [1, 5]]], [:@int, "2", [1, 7]]]]]]
...
ruby-2.4.9          [:program,
                     [[:massign,
                       [:mlhs_add_star, [[:@ident, "a", [1, 0]]], [:@ident, "b", [1, 3]]],
                       [:mrhs_new_from_args, [[:@int, "1", [1, 5]]], [:@int, "2", [1, 7]]]]]]
ruby-2.5.0-preview1 [:program,
                     [[:massign,
                       [[:var_field, [:@ident, "a", [1, 0]]],
                        [:rest_param, [:var_field, [:@ident, "b", [1, 3]]]]],
                       [:mrhs_new_from_args, [[:@int, "1", [1, 5]]], [:@int, "2", [1, 7]]]]]]
...
ruby-2.7.0-rc2      [:program,
                     [[:massign,
                       [[:var_field, [:@ident, "a", [1, 0]]],
                        [:rest_param, [:var_field, [:@ident, "b", [1, 3]]]]],
                       [:mrhs_new_from_args, [[:@int, "1", [1, 5]]], [:@int, "2", [1, 7]]]]]]
```

```
$ docker run -it --rm rubylang/all-ruby env ALL_RUBY_SINCE=ruby-2.4 ./all-ruby -rpp -rripper -e 'pp Ripper.sexp("a,b=1,2")'
ruby-2.4.0          [:program,
                     [[:massign,
                       [[:@ident, "a", [1, 0]], [:@ident, "b", [1, 2]]],
                       [:mrhs_new_from_args, [[:@int, "1", [1, 4]]], [:@int, "2", [1, 6]]]]]]
...
ruby-2.4.9          [:program,
                     [[:massign,
                       [[:@ident, "a", [1, 0]], [:@ident, "b", [1, 2]]],
                       [:mrhs_new_from_args, [[:@int, "1", [1, 4]]], [:@int, "2", [1, 6]]]]]]
ruby-2.5.0-preview1 [:program,
                     [[:massign,
                       [[:var_field, [:@ident, "a", [1, 0]]],
                        [:var_field, [:@ident, "b", [1, 2]]]],
                       [:mrhs_new_from_args, [[:@int, "1", [1, 4]]], [:@int, "2", [1, 6]]]]]]
...
ruby-2.7.0-preview1 [:program,
                     [[:massign,
                       [[:var_field, [:@ident, "a", [1, 0]]],
                        [:var_field, [:@ident, "b", [1, 2]]]],
                       [:mrhs_new_from_args, [[:@int, "1", [1, 4]]], [:@int, "2", [1, 6]]]]]]
```
